refactor(mpc/domain): Create authenticator class, make index.ts an entrypoint for import/export only
refactor(stackup/nextJs): Rename mpc sdk to mpc authenticator

refactor(stackup/cli): Rename mpc sdk to mpc authenticator

refactor(biconomy/cli): Rename mpc sdk to mpc authenticator

refactor(biconomy/nextJs): Rename mpc sdk to mpc authenticator
