+++ packages/biconomy/nextJs/src/app/(client)/(setup)/backup/page.tsx
@@ -10,17 +10,17 @@ import { WALLET_STATUS } from "@/constants";
 import { RouteLoader } from "@/components/routeLoader";
 import { layoutClassName } from "@/utils/ui";
 import { getPairingStatus, setPairingStatus } from "@/storage/localStorage";
-import { useMpcSdk } from "@/hooks/useMpcSdk";
+import { useMpcAuth } from "@/hooks/useMpcAuth";
 
 function Page() {
     const router = useRouter();
-    const mpcSdk = useMpcSdk();
+    const mpcAuth = useMpcAuth();
     const moveToNext = () => {
         setPairingStatus(WALLET_STATUS.BackedUp);
         router.replace("/afterBackup");
     };
     const status = getPairingStatus();
-    if (status !== WALLET_STATUS.Paired || mpcSdk.accountManager.isPasswordReady()) {
+    if (status !== WALLET_STATUS.Paired || mpcAuth.accountManager.isPasswordReady()) {
         return <RouteLoader />;
     }
     return (
