+++ packages/biconomy/nextJs/src/app/(client)/(setup)/pair/page.tsx
@@ -23,10 +23,10 @@ import {
     setPairingStatus,
 } from "@/storage/localStorage";
 import { PairingSessionData } from "@silencelaboratories/mpc-sdk/lib/esm/types";
-import { useMpcSdk } from "@/hooks/useMpcSdk";
+import { useMpcAuth } from "@/hooks/useMpcAuth";
 
 function Page() {
-    const mpcSdk = useMpcSdk();
+    const mpcAuth = useMpcAuth();
     const router = useRouter();
     const query = useSearchParams();
     const isRepairing = query.get("repair");
@@ -49,12 +49,12 @@ function Page() {
         password: string
     ) => {
         try {
-            await mpcSdk.runEndPairingSession(
+            await mpcAuth.runEndPairingSession(
                 pairingSessionData,
                 oldEoa ?? undefined,
                 password
             );
-            const eoa = mpcSdk.accountManager.getEoa();
+            const eoa = mpcAuth.accountManager.getEoa();
             if (isRepairing && oldEoa !== null && eoa !== oldEoa) {
                 setPairingStatus(WALLET_STATUS.Mismatched);
                 router.replace("/mismatchAccounts");
@@ -72,13 +72,13 @@ function Page() {
     const generateWallet = async () => {
         (async () => {
             try {
-                const qrCode = await mpcSdk.initPairing();
+                const qrCode = await mpcAuth.initPairing();
                 setQr(qrCode);
                 setSeconds(MAX_SECONDS);
                 setEnterPwSeconds(MAX_ENTER_PW_SECONDS);
 
                 const pairingSessionData =
-                    await mpcSdk.runStartPairingSession();
+                    await mpcAuth.runStartPairingSession();
 
                 // QR is scanned
                 setLoading(true);
@@ -87,11 +87,11 @@ function Page() {
                     setPairingSessionDataState(pairingSessionData);
                     setShowPasswordScreen(true);
                 } else {
-                    await mpcSdk.runEndPairingSession(
+                    await mpcAuth.runEndPairingSession(
                         pairingSessionData,
                         oldEoa ?? undefined,
                     );
-                    await mpcSdk.runKeygen();
+                    await mpcAuth.runKeygen();
 
                     setPairingStatus(WALLET_STATUS.Paired);
                     router.replace("/backup");
