+++ packages/stackup/cli/README.md
@@ -37,9 +37,11 @@ To set up the rpcUrl, create an instance at StackUp:
 2. Fill in your Base urls and API keys in `.env`:
 
     - Replace `your_stackup_api_key_here` with your actual stackup's API key.
+    - To run in development environment, add `NODE_ENV=development`.
 
         ```env
         API_KEY=your_stsackup_api_key_here
+        NODE_ENV=development
         ```
 
 ## Step 3: Install Dependencies
