--- packages/stackup/cli/config.json
@@ -1,43 +1,0 @@
-{
-  "rpcUrl": "https://api.stackup.sh/v1/node/32bbc56086c93278c34d5b3376a487e6b57147f052ec41688c1ad65bd984af7e",
-  "paymaster": {
-    "rpcUrl": "https://api.stackup.sh/v1/paymasterhttps://api.stackup.sh/v1/paymaster/32bbc56086c93278c34d5b3376a487e6b57147f052ec41688c1ad65bd984af7e",
-    "context": {}
-  },
-  "_isSigner": true,
-  "address": "0xAC5B08F56FBFEd74f2de4990d5db4f244A0dd465",
-  "public_key": "5e76399839ae5433d4a62abd1e0e9c6195417a6bcef729a6c9877a405cc3fd2f7af5d19707e869d3d89e8dd015d649a1065357e9c065e295b9365560e81fe5cc",
-  "p1KeyShare": {
-    "x1": "7d5ab7efd87abe2e949b52b414ff953ef5d4fea58d04f0ed9905083bbb46f965",
-    "public_key": "5e76399839ae5433d4a62abd1e0e9c6195417a6bcef729a6c9877a405cc3fd2f7af5d19707e869d3d89e8dd015d649a1065357e9c065e295b9365560e81fe5cc",
-    "paillier_private_key": {
-      "p": "AZI1G5KRM9Bbhmy1FNG6lWMH/mlBwhaV84hb5+mzWzUCKP0hqjSisOxy/yasnE7SK4q+d7kk4dqz3SypxL4EakxgIaMHCj1fyc+v3HNzpXh89kqv66IRBgNSavjUdPfLY0dnLOvW6XAIp84OClFG+cfRlPPlanWX2FipvA9tDioj",
-      "q": "kEBFrZuCbaac40Keo0ivB4Zo5ZL5wZ7wXrr5vMPZZx4D6cbJVSDoNCWCkiwFacI4n3QCu/Ioxa7NCzqiq1G13Ev6FpQrO8GnhSA50SR5ggkRpCCuKTOQRUeT8LQPJUsWbu8gMIi9ng0I/qE7OgvKDiA61GqfP6K9EjHdpxo0oVE="
-    },
-    "paillier_public_key": "4qLaQmj2qEbJrgkCQX+q/cqshPhkwRkcg0+39V0FfRslcvhGcVuEJphZXVCEtTSRy1igoS1l05dDOVB7H3a5BwpuMyda9IJI0e4iNEuBSdcPBNzgC3nzVB5TW8qBlqJYSSS4bq49KcCCq49f7Nt9QCZuCuQ+EB2zR+KV3p5Ee7p/bxy/XEk150HsUvLF1Hpca99eX1RYUzOay1Yq20XgVWGKcKTZ/keOUDh03aBNXyHSzIJ/EHTf/S7fYCy8UgY/ejSgseSh1Q3uku+wK73Wa4oZeV8CRKn2QJB8/FGMCai6B1YbKGT3pHFAki3ybH5Mm+hqyH0VeoN2gH90hxdYEw=="
-  },
-  "keygenResult": {
-    "distributedKey": {
-      "publicKey": "5e76399839ae5433d4a62abd1e0e9c6195417a6bcef729a6c9877a405cc3fd2f7af5d19707e869d3d89e8dd015d649a1065357e9c065e295b9365560e81fe5cc",
-      "accountId": 1,
-      "keyShareData": {
-        "x1": "7d5ab7efd87abe2e949b52b414ff953ef5d4fea58d04f0ed9905083bbb46f965",
-        "public_key": "5e76399839ae5433d4a62abd1e0e9c6195417a6bcef729a6c9877a405cc3fd2f7af5d19707e869d3d89e8dd015d649a1065357e9c065e295b9365560e81fe5cc",
-        "paillier_private_key": {
-          "p": "AZI1G5KRM9Bbhmy1FNG6lWMH/mlBwhaV84hb5+mzWzUCKP0hqjSisOxy/yasnE7SK4q+d7kk4dqz3SypxL4EakxgIaMHCj1fyc+v3HNzpXh89kqv66IRBgNSavjUdPfLY0dnLOvW6XAIp84OClFG+cfRlPPlanWX2FipvA9tDioj",
-          "q": "kEBFrZuCbaac40Keo0ivB4Zo5ZL5wZ7wXrr5vMPZZx4D6cbJVSDoNCWCkiwFacI4n3QCu/Ioxa7NCzqiq1G13Ev6FpQrO8GnhSA50SR5ggkRpCCuKTOQRUeT8LQPJUsWbu8gMIi9ng0I/qE7OgvKDiA61GqfP6K9EjHdpxo0oVE="
-        },
-        "paillier_public_key": "4qLaQmj2qEbJrgkCQX+q/cqshPhkwRkcg0+39V0FfRslcvhGcVuEJphZXVCEtTSRy1igoS1l05dDOVB7H3a5BwpuMyda9IJI0e4iNEuBSdcPBNzgC3nzVB5TW8qBlqJYSSS4bq49KcCCq49f7Nt9QCZuCuQ+EB2zR+KV3p5Ee7p/bxy/XEk150HsUvLF1Hpca99eX1RYUzOay1Yq20XgVWGKcKTZ/keOUDh03aBNXyHSzIJ/EHTf/S7fYCy8UgY/ejSgseSh1Q3uku+wK73Wa4oZeV8CRKn2QJB8/FGMCai6B1YbKGT3pHFAki3ybH5Mm+hqyH0VeoN2gH90hxdYEw=="
-      }
-    },
-    "elapsedTime": 2508
-  },
-  "mpcSdk": {
-    "TOKEN_LIFE_TIME": 60000,
-    "walletId": "stackup",
-    "storage": {},
-    "accountManager": {
-      "storage": {}
-    }
-  }
-}
