+++ packages/stackup/cli/scripts/simpleAccount/address.ts
@@ -5,7 +5,7 @@ import { Presets } from "userop";
 // @ts-ignore
 import config from "../../config.json";
 import chalk from "chalk";
-import { mpcSdk } from "../../mpc";
+import { mpcAuth } from "../../mpc";
 import { MpcSigner } from "@silencelaboratories/mpc-sdk/lib/cjs/domain/signer";
 
 export default async function main() {
@@ -15,7 +15,7 @@ export default async function main() {
       config.public_key,
       config.p1KeyShare,
       config.keygenResult,
-      mpcSdk
+      mpcAuth
     ),
     config.rpcUrl
   );
