+++ packages/stackup/nextJs/src/app/(client)/homescreen/page.tsx
@@ -20,7 +20,7 @@ import Image from "next/image";
 import { SEPOLIA, WALLET_STATUS } from "@/constants";
 import Footer from "@/components/footer";
 import { RouteLoader } from "@/components/routeLoader";
-import { useMpcSdk } from "@/hooks/useMpcSdk";
+import { useMpcAuth } from "@/hooks/useMpcAuth";
 import { AccountData } from "@silencelaboratories/mpc-sdk/lib/esm/types";
 import {
     clearOldEoa,
@@ -31,7 +31,7 @@ import {
 } from "@/storage/localStorage";
 
 const Homescreen: React.FC = () => {
-    const mpcSdk = useMpcSdk();
+    const mpcAuth = useMpcAuth();
     const oldEoa = getOldEoa();
     const router = useRouter();
     const [walletAccount, setWalletAccount] = useState<AccountData>();
@@ -50,7 +50,7 @@ const Homescreen: React.FC = () => {
 
     useEffect(() => {
         try {
-            setIsPasswordReady(mpcSdk.accountManager.isPasswordReady());
+            setIsPasswordReady(mpcAuth.accountManager.isPasswordReady());
         } catch (error) {
             console.error("isPasswordReady error", error);
         }
@@ -58,14 +58,14 @@ const Homescreen: React.FC = () => {
 
     useEffect(() => {
         try {
-            const account = mpcSdk.accountManager.getSmartContractAccount();
+            const account = mpcAuth.accountManager.getSmartContractAccount();
             if (!account) {
                 setPairingStatus(WALLET_STATUS.BackedUp);
                 router.replace("/mint");
                 return;
             }
 
-            const eoa = mpcSdk.accountManager.getEoa();
+            const eoa = mpcAuth.accountManager.getEoa();
             if (!eoa) {
                 setPairingStatus(WALLET_STATUS.Unpaired);
                 router.replace("/intro");
@@ -265,7 +265,7 @@ const Homescreen: React.FC = () => {
                 const result = await sendTransaction(
                     recipientAddress,
                     amount,
-                    mpcSdk
+                    mpcAuth
                 );
                 if (!result.transactionHash) {
                     setShowTransactionInitiatedBanner(false);
@@ -289,7 +289,7 @@ const Homescreen: React.FC = () => {
 
     const logout = (event: React.MouseEvent): void => {
         event.preventDefault();
-        mpcSdk.signOut();
+        mpcAuth.signOut();
         clearOldEoa();
         setPairingStatus(WALLET_STATUS.Unpaired);
         router.push("/intro");
